Make button states react to selection changes
Show what kind of zooming is possible by enabling or disabling the
zoom buttons based on the type of the selection that has been done.

Change in button state happens whenever the selection is changed and
is implemented by extending the MandelbrotPanelListener interface to
contain the method selectionUpdate(MandelbrotPanel.SelectionState
state), where state indicates the current selection state.