Improve selection calculations
Previous code used the complex coordinates for calculation of center,
width, and height of new selections.  This improved code also makes
use of the screen coordinates of the mouse positions in those
calculations.  These are represented as integers and doesn't suffer
from the same loss of precision as previous code.